pyproject.toml
[project]
name = "publication-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.7",
    "djangorestframework-simplejwt>=5.5.1",
    "djangorestframework>=3.16.1",
    "dotenv>=0.9.9",
    "drf-spectacular>=0.28.0",
    "loadenv>=0.1.1",
    "psycopg2>=2.9.11",
]
